Modify rules to eliminate ambiguities and make text more concise
- Make writing more concise in goal kicks, corner kicks and throw-ins by
referring to free kick section
- Change the rule in direct free kicks: only goals on opposing teams can
be scored (strange, but those are FIFA rules). With this change, corner
kicks and goal kicks are identical to direct free kicks
- Remove unnecessary and contradictory "A free kick is taken from where
the offence occurred" from fouls and misconduct
- Changed wording in referee.tex: "repositions the ball to a neutral
position..." to "places the ball as specified by the appropriate law"
- start-and-restart-of-play.tex: Changed subsections to subsubsections,
which is what they should be. Eliminated  redundant "Special
Circumstances" subsection