Remove the free ticket for rule violations
When a team has failed to place the ball five times in a row, all
free kicks were awarded to the opposing team. This team had a free
ticket for rule violations, as it always got a free kick for it.
Now, it is only awarded a free kick if the ball left the field